Demo tweaks.
