+++ .circleci/.env.ci
@@ -1,21 +1,30 @@
-OPTIMIZE=3
+OPTIMIZE=2
+BROTLI=0
+GZIP=0
+
+WITH_BCMATH=1
+WITH_CALENDAR=1
+WITH_CTYPE=1
+WITH_EXIF=1
+WITH_FILTER=1
+WITH_MBSTRING=1
+WITH_PIB=1
+WITH_PHAR=1
+WITH_TOKENIZER=1
 
 WITH_ICONV=1
 WITH_LIBXML=1
 WITH_LIBZIP=1
-WITH_OPENSSL=1
 WITH_SQLITE=1
 WITH_VRZNO=1
-
 WITH_ZLIB=1
+
+WITH_OPENSSL=1
 WITH_GD=1
-WITH_LIBJPEG=1
 WITH_LIBPNG=1
+WITH_LIBJPEG=1
 WITH_FREETYPE=1
 
 WITH_ICU=1
 WITH_TIDY=1
 WITH_EXIF=1
-
-GZIP=0
-BROTLI=0
