+++ demo-source/src/Confirm.js
@@ -4,7 +4,7 @@ export default function Confirm({message, onConfirm, onCancel})
 {
 	return (
 		<div className="Confirm">
-			<div className="dialog bevel column padded">
+			<div className="dialog bevel column">
 				<span className='inset padded'>{message}</span>
 				<div className="right">
 					<button className='padded' onClick = {onConfirm}>Confirm</button>
