+++ demo-source/src/EditorEntry.css
@@ -23,7 +23,7 @@
 }
 
 .editor-entry .icon {
-	height: 1.5rem;
+	height: 1.15rem;
 	vertical-align: middle;
 	margin-right: 0.5rem;
 }
