+++ demo-source/src/SelectFramework.css
@@ -8,7 +8,6 @@
 }
 
 .select-framework .frameworks {
-	padding: 0.5rem;
 	margin: 0.5rem;
 }
