+++ packages/zlib/test/basic.mjs
@@ -1,0 +1,20 @@
+import { test } from 'node:test';
+import { strict as assert } from 'node:assert';
+import { PhpNode } from '../../../packages/php-wasm/PhpNode.mjs';
+
+test('Zlib Extension is enabled.', async () => {
+	const php = new PhpNode();
+
+	let stdOut = '', stdErr = '';
+
+	php.addEventListener('output', (event) => event.detail.forEach(line => void (stdOut += line)));
+	php.addEventListener('error',  (event) => event.detail.forEach(line => void (stdErr += line)));
+
+	await php.binary;
+
+	const exitCode = await php.run(`<?php var_dump(extension_loaded('zlib'));`);
+
+	assert.equal(exitCode, 0);
+	assert.equal(stdOut, `bool(true)\n`);
+	assert.equal(stdErr, '');
+});
