+++ source/PhpWorker.js
@@ -1,6 +1,6 @@
 import { PhpBase } from './PhpBase';
-import { commitTransaction, startTransaction } from './webTransactions';
 import PhpBinary from './php-worker';
+import { commitTransaction, startTransaction } from './webTransactions';
 
 export class PhpWorker extends PhpBase
 {
@@ -17,5 +17,33 @@ export class PhpWorker extends PhpBase
 	commitTransaction()
 	{
 		return commitTransaction(this);
+	}
+
+	async _enqueue(callback, params = [])
+	{
+		let accept, reject;
+
+		const coordinator = new Promise((a,r) => [accept, reject] = [a, r]);
+
+		this.queue.push([callback, params, accept, reject]);
+
+		await navigator.locks.request('php-wasm-fs-lock', async () => {
+			if(!this.queue.length)
+			{
+				return;
+			}
+
+			await (this.autoTransaction ? this.startTransaction() : Promise.resolve());
+
+			while(this.queue.length)
+			{
+				const [callback, params, accept, reject] = this.queue.shift();
+				await callback(...params).then(accept).catch(reject);
+			}
+
+			await (this.autoTransaction ? this.commitTransaction() : Promise.resolve());
+		});
+
+		return coordinator;
 	}
 }
